Removed bijection between Track and Type
This required a new class (TrackWithArray, and the array is actually a PriorityQueue), some casting to and from Track, a pointless implementation of a Parcelizable thingamajig to stop Android Studio from complaining, a database schema modification, and some additional logic in the parser. I'm amazed everything worked without any compile errors or debugging.